feat: Home Page 내 API 적용(#12)
